requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0